Rename isRepoReady file to checkIsRepoReady
