+++ processor/src/checkIsRepoReady.js
